+++ Frontend/src/pages/about.jsx
@@ -2,7 +2,6 @@ import React from 'react'
 import BlogBanner from '../sections/BlogBanner'
 import BlogSection from '../sections/BlogSection'
 import FeaturesSection from '../sections/FeaturesSection'
-import Footer from '../components/Footer'
 import { usePageMeta } from '../shared/lib/usePageMeta.js';
 
 
@@ -13,8 +12,6 @@ function About() {
       <BlogBanner />
       <BlogSection />
       <FeaturesSection />
-      <Footer />
-      
     </div>
   )
 }