+++ Frontend/src/pages/shop.jsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import ProductGrid from '../components/ProductGrid'
 import FeaturesSection from '../sections/FeaturesSection'
-import Footer from '../components/Footer'
 import ShopBanner from '../sections/ShopBanner'
 import { usePageMeta } from '../shared/lib/usePageMeta.js';
 
@@ -12,7 +11,6 @@ function Shop() {
       <ShopBanner/>
       <ProductGrid />
       <FeaturesSection />
-      <Footer />
     </div>
   )
 }